HSTTWO-2250 fix classcastexception for integer

+++ site-toolkit/components/core/src/main/java/org/hippoecm/hst/configuration/channel/ChannelPropertyMapper.java
@@ -248,7 +248,7 @@ public class ChannelPropertyMapper {
                 log.warn("Cannot store a Integer (Long in jcr) '{}' for '{}'. Store default value instead", value, propDef.getName());
                 return defaultValueToJcr(vf, propDef);
             } else {
-                return vf.createValue((Long) value);
+                return vf.createValue(((Integer)value).longValue());
             }
         } else if (value instanceof Long) {
             if (propDef.getValueType() != HstValueType.LONG) {
